+++ src/steps/storage/index.ts
@@ -1,4 +1,11 @@
-import { IntegrationStep } from '@jupiterone/integration-sdk-core';
+import {
+  createDirectRelationship,
+  Entity,
+  generateRelationshipType,
+  getRawData,
+  IntegrationStep,
+  RelationshipClass,
+} from '@jupiterone/integration-sdk-core';
 import { CloudStorageClient } from './client';
 import { IntegrationConfig, IntegrationStepContext } from '../../types';
 import { createCloudStorageBucketEntity } from './converters';
@@ -6,10 +13,12 @@ import {
   CLOUD_STORAGE_BUCKET_ENTITY_TYPE,
   STEP_CLOUD_STORAGE_BUCKETS,
   CLOUD_STORAGE_BUCKET_ENTITY_CLASS,
+  CREATE_PROJECT_BUCKET_RELATIONSHIPS,
 } from './constants';
 import { storage_v1 } from 'googleapis';
 import { isMemberPublic } from '../../utils/iam';
 import { publishUnprocessedBucketsEvent } from '../../utils/events';
+import { PROJECT_ENTITY_TYPE } from '../resource-manager';
 
 export * from './constants';
 
@@ -84,6 +93,43 @@ export async function fetchStorageBuckets(
   }
 }
 
+export async function createProjectStorageBucketRelationships(
+  context: IntegrationStepContext,
+): Promise<void> {
+  const { jobState, logger } = context;
+  await jobState.iterateEntities(
+    { _type: CLOUD_STORAGE_BUCKET_ENTITY_TYPE },
+    async (bucketEntity: Entity) => {
+      const projectNumber =
+        getRawData<storage_v1.Schema$Bucket>(bucketEntity)?.projectNumber;
+      if (typeof projectNumber !== 'string') {
+        logger.warn(
+          { bucketEntityKey: bucketEntity._key },
+          'Bucket does not have a projectNumber',
+        );
+        return;
+      }
+      const projectEntity = await jobState.findEntity(
+        'projects/' + projectNumber,
+      );
+      if (!projectEntity) {
+        logger.warn(
+          { projectNumber },
+          'Unable to find project entity in jobState',
+        );
+        return;
+      }
+      await jobState.addRelationship(
+        createDirectRelationship({
+          _class: RelationshipClass.HAS,
+          from: projectEntity,
+          to: bucketEntity,
+        }),
+      );
+    },
+  );
+}
+
 export const storageSteps: IntegrationStep<IntegrationConfig>[] = [
   {
     id: STEP_CLOUD_STORAGE_BUCKETS,
@@ -97,5 +143,25 @@ export const storageSteps: IntegrationStep<IntegrationConfig>[] = [
     ],
     relationships: [],
     executionHandler: fetchStorageBuckets,
+  },
+  {
+    id: CREATE_PROJECT_BUCKET_RELATIONSHIPS,
+    name: 'Create Project Storage Bucket Relationships',
+    entities: [],
+    relationships: [
+      {
+        _class: RelationshipClass.HAS,
+        _type: generateRelationshipType(
+          RelationshipClass.HAS,
+          PROJECT_ENTITY_TYPE,
+          CLOUD_STORAGE_BUCKET_ENTITY_TYPE,
+        ),
+        sourceType: PROJECT_ENTITY_TYPE,
+        targetType: CLOUD_STORAGE_BUCKET_ENTITY_TYPE,
+      },
+    ],
+    dependsOn: [],
+    executionHandler: createProjectStorageBucketRelationships,
+    dependencyGraphId: 'last', // Needs to be last in order to ensure there are projects in the jobState
   },
 ];